requirements.txt
openai
requests
beautifulsoup4
PyYAML
pandas
tqdm
python-dateutil
PyPDF2
langdetect
PyMuPDF
